requirements_datadrift.txt
anyio
pandas
evidently
numpy
Pillow
fastapi[standard]
fastapi_utils
google-cloud-storage
